+++ core/src/main/java/Gameplay/Movimientos/Melee/Aranazo.java
@@ -10,7 +10,7 @@ public class Aranazo extends MovimientoMelee {
 
     public Aranazo(GestorColisiones gestorColisiones) {
         super("Arañazo", GestorAssets.get(Constantes.PNG_2, Texture.class), 10f, 40f,
-            10, 550f, gestorColisiones);
+            10, 50f, gestorColisiones);
     }
 
 }